--- a/src/pages/GearPage.tsx
+++ b/src/pages/GearPage.tsx
@@ -191,7 +191,7 @@ export default function GearPage() {
       
       <main className="bg-gray-50 dark:bg-gray-900 min-h-screen">
         {/* Hero Section */}
-        <section className="bg-gradient-to-br from-blue-600 via-purple-600 to-blue-800 dark:from-blue-800 dark:via-purple-800 dark:to-blue-900 text-white py-20">
+        <section className="bg-linear-to-br from-blue-600 via-purple-600 to-blue-800 dark:from-blue-800 dark:via-purple-800 dark:to-blue-900 text-white py-20">
           <div className="container mx-auto px-4 text-center">
             <h1 className="text-4xl md:text-5xl font-bold mb-6 drop-shadow-lg">
               My Gear