… package.json
… pnpm-lock.yaml
… tailwind.config.js